refactored bug report
